requirements.txt
streamlit>=1.25.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.1
chromadb>=0.3.25
langchain>=0.0.320
deep-translator>=1.8.1
pdfplumber>=0.7.6
python-docx>=0.8.11
diskcache>=5.4.0
scikit-learn>=1.2.2
joblib>=1.2.0
matplotlib>=3.7.1
numpy>=1.24.0
requests>=2.31.0
diff-match-patch>=20200713
fastapi>=0.103.0
uvicorn>=0.23.0
